ai-worker/requirements.txt
grpcio>=1.69.0
grpcio-reflection>=1.69.0
grpcio-health-checking>=1.69.0
protobuf>=5.29.0
pydantic-settings>=2.7.0
structlog>=24.4.0
